+++ Source/Objects/Hardware/SBC/SBC_Code/SBC_Config.h
@@ -2,6 +2,7 @@
 #ifndef _H_SBC_CONFIG_
 #define _H_SBC_CONFIG_
 #include <sys/types.h>
+#include <stdint.h>
 //-----------------------------------------------------------
 //This program was prepared for the Regents of the University of 
 //Washington at the Center for Experimental Nuclear Physics and 
@@ -17,45 +18,45 @@
 //-------------------------------------------------------------
 
 // ---------------------------------------------------------------------------- 
-//	Generic hardware configuration structure used by both Mac and eCPU code.
-#define MAX_CARDS			20
+//    Generic hardware configuration structure used by both Mac and eCPU code.
+#define MAX_CARDS            20
 
 typedef struct {
-	unsigned long header;
-    unsigned long total_cards;					// total sum of all cards
-    struct {									// structure required for card
-        unsigned long	hw_type_id;				// unique hardware identifier code
-        unsigned long	hw_mask[10];			// hardware identifier mask to OR into data word
-		unsigned long	slot;					// slot identifier
-		unsigned long	crate;					// crate identifier
-		unsigned long	base_add;				// base addresses for each card
-		unsigned long   add_mod;				// address modifier (if needed)
-		unsigned long	deviceSpecificData[5];	// a card can use this block as needed.
-		unsigned long	next_Card_Index;		// next card_info index to be read after this one.		
-		unsigned long 	num_Trigger_Indexes;	// number of triggers for this card
-		unsigned long	next_Trigger_Index[3];	//card_info index for device specific trigger
-	} card_info[MAX_CARDS];
+    uint32_t header;
+    uint32_t total_cards;                    // total sum of all cards
+    struct {                                    // structure required for card
+        uint32_t hw_type_id;                // unique hardware identifier code
+        uint32_t hw_mask[10];            // hardware identifier mask to OR into data word
+        uint32_t slot;                    // slot identifier
+        uint32_t crate;                    // crate identifier
+        uint32_t base_add;                // base addresses for each card
+        uint32_t add_mod;                // address modifier (if needed)
+        uint32_t deviceSpecificData[5];    // a card can use this block as needed.
+        uint32_t next_Card_Index;        // next card_info index to be read after this one.        
+        uint32_t num_Trigger_Indexes;    // number of triggers for this card
+        uint32_t next_Trigger_Index[3];    //card_info index for device specific trigger
+    } card_info[MAX_CARDS];
 } SBC_crate_config;
 
-#define kSBC_CrateConfigSizeLongs sizeof(SBC_crate_config)/sizeof(unsigned long)
+#define kSBC_CrateConfigSizeLongs sizeof(SBC_crate_config)/sizeof(uint32_t)
 #define kSBC_CrateConfigSizeBytes sizeof(SBC_crate_config)
 
 typedef struct {
-	unsigned long statusBits;
-	unsigned long readCycles;
-	unsigned long bufferSize;
-	unsigned long readIndex;
-	unsigned long writeIndex;
-	unsigned long lostByteCount;
-	unsigned long amountInBuffer;
-	unsigned long recordsTransfered;
-	unsigned long wrapArounds;
+    uint32_t statusBits;
+    uint32_t readCycles;
+    uint32_t bufferSize;
+    uint32_t readIndex;
+    uint32_t writeIndex;
+    uint32_t lostByteCount;
+    uint32_t amountInBuffer;
+    uint32_t recordsTransfered;
+    uint32_t wrapArounds;
 } SBC_info_struct;
 
-#define kSBC_ConfigLoadedMask	(0x1 << 0)
-#define kSBC_RunningMask		(0x1 << 1)
+#define kSBC_ConfigLoadedMask    (0x1 << 0)
+#define kSBC_RunningMask        (0x1 << 1)
 
-#define kSBC_InfoStructSizeLongs sizeof(SBC_info_struct)/sizeof(unsigned long)
+#define kSBC_InfoStructSizeLongs sizeof(SBC_info_struct)/sizeof(uint32_t)
 #define kSBC_InfoStructSizeBytes sizeof(SBC_info_struct)
 
 #endif
